requirements.txt
discord.py
bs4
requests
html2text
